Eliminate previous ci tests. Merge ci keyring tests to unittests, the other ci tests just duplicate other unit tests.
Change Makefile to just run grunt and npm tasks.
