--- test/ci-tests-all.js
@@ -1,3 +1,0 @@
-require('./ci-quick.js');
-require('./general');
-require('./crypto');
