--- test/test-all.js
@@ -1,20 +1,0 @@
-
-module.exports = require('./unit.js');
-
-require('./crypto/hash/sha.js');
-require('./crypto/hash/md5.js');
-require('./crypto/hash/ripemd.js');
-
-require('./crypto/cipher/aes.js');
-require('./crypto/cipher/blowfish.js');
-require('./crypto/cipher/cast5.js');
-require('./crypto/cipher/des.js');
-require('./crypto/cipher/twofish.js');
-
-require('./crypto/crypto.js');
-
-require('./general/basic.js');
-require('./general/key.js');
-require('./general/keyring.js');
-require('./general/packet.js');
-require('./general/signature.js');
