--- test/test.js
@@ -1,6 +1,0 @@
-
-module.exports = require('./unit.js');
-
-require('./test-all.js');
-
-module.exports.run_all();
