+++ test/unittests.js
@@ -1,0 +1,4 @@
+describe('Unit Tests', function () {
+  require('./general');
+  require('./crypto');
+});
